+++ demos/manual/src/main/java/is/codion/framework/demos/manual/store/ui/CustomerAddressEditPanel.java
@@ -38,7 +38,7 @@ public class CustomerAddressEditPanel extends EntityEditPanel {
 
 		createForeignKeyComboBoxPanel(CustomerAddress.ADDRESS_FK, this::createAddressEditPanel)
 						.preferredWidth(280)
-						.add(true);
+						.addButton(true);
 
 		setLayout(borderLayout());